Layout rewrite continues
Added missing front matter to each page and to the main SCSS, so that Jekyll catches them all when building locally.

Sidebar now uses relative URLs for its hyperlinks.

Sidebar pieces now use H1, as they are sectioning elements and heading numbers should be local to the containing sectioning element.

Moved the header element into the main element.

Added a skip-to-content link to the top of the page, for screen readers.

Extensive rewrites to the CSS, taking advantage of SASS nesting. Disabled the original Jekyll theme since its layout code was counterproductive for our purposes; recreated font settings using rem instead of px. Additionally, some style properties are configurable as SCSS variables (this is not comprehensive).

Marked the "jekyll-seo-tag" plug-in as required; GitHub Pages documentation doesn't list it, but that does seem to be the plug-in which powers the Jekyll `seo` tag, which we use.

Removed non-useful alt text (just the list name and version number) from the screenshots in the main page.