Testimonials; troubleshooting
Work on including testimonials as part of this site (with the issue still hosted on the other repo) rather than as a separate site, so that the layout can be kept unified.

GitHub Pages is doing some really, really nasty stuff to the CSS -- stuff that isn't happening when I test Jekyll locally.